implementing document-fetcher and splitting xhr-impl into a base class
